+++ tests/test_data_proxy.py
@@ -49,13 +49,13 @@ class TestDataProxy(BaseTest):
 
         MyDataProxy = data_proxy_factory('My', MySchema())
         d = MyDataProxy()
-        d.load({'a': 1, 'b': 2})
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
         assert d.to_mongo() == {'a': 1, 'in_mongo_b': 2}
 
         d.set('a', 3)
         assert d.to_mongo(update=True) == {'$set': {'a': 3}}
 
-        d.load({'a': 4, 'b': 5})
+        d.from_mongo({'a': 4, 'in_mongo_b': 5})
         assert d.to_mongo(update=True) is None
         assert d.to_mongo() == {'a': 4, 'in_mongo_b': 5}
 
@@ -70,13 +70,18 @@ class TestDataProxy(BaseTest):
 
         MyDataProxy = data_proxy_factory('My', MySchema())
         d = MyDataProxy()
+        assert d.get_modified_fields() == []
         d.load({'a': 1, 'b': 2})
+        assert list(sorted(d.get_modified_fields())) == ['a', 'b']
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
+        assert d.get_modified_fields() == []
         assert d.to_mongo() == {'a': 1, 'in_mongo_b': 2}
         assert d.to_mongo(update=True) is None
         d.set('a', 3)
         d.delete('b')
         assert d.to_mongo(update=True) == {'$set': {'a': 3}, '$unset': {'in_mongo_b': ''}}
         d.clear_modified()
+        assert d.get_modified_fields() == []
         assert d.to_mongo(update=True) is None
         assert d.to_mongo() == {'a': 3}
 
@@ -111,12 +116,12 @@ class TestDataProxy(BaseTest):
 
         MyDataProxy = data_proxy_factory('My', MySchema())
         d = MyDataProxy()
-        d.load({'a': 1, 'b': 2})
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
         d.set('a', 3)
         assert d.to_mongo() == {'a': 3, 'in_mongo_b': 2}
         assert d.to_mongo(update=True) == {'$set': {'a': 3}}
 
-        d.load({'a': 1, 'b': 2})
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
         d.set('b', 3)
         assert d.to_mongo() == {'a': 1, 'in_mongo_b': 3}
         assert d.to_mongo(update=True) == {'$set': {'in_mongo_b': 3}}
@@ -132,7 +137,7 @@ class TestDataProxy(BaseTest):
 
         MyDataProxy = data_proxy_factory('My', MySchema())
         d = MyDataProxy()
-        d.load({'a': 1, 'b': 2})
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
         d.delete('b')
         assert d.to_mongo() == {'a': 1}
         assert d.to_mongo(update=True) == {'$unset': {'in_mongo_b': ''}}
@@ -210,7 +215,8 @@ class TestDataProxy(BaseTest):
             b = fields.IntField(attribute='in_mongo_b')
 
         MyDataProxy = data_proxy_factory('My', MySchema())
-        d = MyDataProxy(data={'a': 1, 'b': 2})
+        d = MyDataProxy()
+        d.from_mongo({'a': 1, 'in_mongo_b': 2})
         assert d.get_by_mongo_name('in_mongo_b') == 2
         assert d.get_by_mongo_name('a') == 1
         with pytest.raises(KeyError):