+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 dev
 ---
 
+* Document&EmbeddedDocument store fields passed during initialization
+  as modified (see #50)
 * Required field inside embedded document are handled correctly (see #61)
 * Document support marshmallow's pre/post processors